convert article-series to tsx]
