--- content/webapp/pages/event-series.js
@@ -1,168 +1,0 @@
-// @flow
-import type { Context } from 'next';
-import type { EventSeries } from '@weco/common/model/event-series';
-import type { UiEvent } from '@weco/common/model/events';
-import { Component } from 'react';
-import { getEventSeries } from '@weco/common/services/prismic/event-series';
-// $FlowFixMe (tsx)
-import PageLayout from '@weco/common/views/components/PageLayout/PageLayout';
-// $FlowFixMe (tsx)
-import ContentPage from '@weco/common/views/components/ContentPage/ContentPage';
-// $FlowFixMe (tsx)
-import Body from '@weco/common/views/components/Body/Body';
-import SearchResults from '@weco/common/views/components/SearchResults/SearchResults';
-// $FlowFixMe(tsx)
-import HeaderBackground from '@weco/common/views/components/HeaderBackground/HeaderBackground';
-import PageHeader, {
-  getFeaturedMedia,
-} from '@weco/common/views/components/PageHeader/PageHeader';
-import { convertImageUri } from '@weco/common/utils/convert-image-uri';
-import { eventLd } from '@weco/common/utils/json-ld';
-// $FlowFixMe
-import { convertJsonToDates } from './event';
-// $FlowFixMe (tsx)
-import Space from '@weco/common/views/components/styled/Space';
-// $FlowFixMe
-import { getGlobalContextData } from '@weco/common/views/components/GlobalContextProvider/GlobalContextProvider';
-
-type Props = {|
-  series: EventSeries,
-  events: UiEvent[],
-  globalContextData: any,
-|};
-
-export class EventSeriesPage extends Component<Props> {
-  static getInitialProps = async (ctx: Context) => {
-    const globalContextData = getGlobalContextData(ctx);
-    const { id, memoizedPrismic } = ctx.query;
-    const seriesAndEvents = await getEventSeries(
-      ctx.req,
-      {
-        id,
-        pageSize: 100,
-      },
-      memoizedPrismic
-    );
-
-    if (seriesAndEvents) {
-      const { series, events } = seriesAndEvents;
-      return {
-        series,
-        events,
-        globalContextData,
-      };
-    } else {
-      return { statusCode: 404 };
-    }
-  };
-
-  render() {
-    const { globalContextData, series } = this.props;
-    const jsonEvents = this.props.events;
-    const events = jsonEvents.map(convertJsonToDates);
-
-    const breadcrumbs = {
-      items: [
-        {
-          url: '/events',
-          text: 'Events',
-        },
-        {
-          url: `/events-series/${series.id}`,
-          text: series.title,
-          isHidden: true,
-        },
-      ],
-    };
-
-    const genericFields = {
-      id: series.id,
-      title: series.title,
-      contributors: series.contributors,
-      contributorsTitle: series.contributorsTitle,
-      promo: series.promo,
-      body: series.body,
-      standfirst: series.standfirst,
-      promoImage: series.promoImage,
-      promoText: series.promoText,
-      image: series.image,
-      squareImage: series.squareImage,
-      widescreenImage: series.widescreenImage,
-      superWidescreenImage: series.superWidescreenImage,
-      labels: series.labels,
-      metadataDescription: series.metadataDescription,
-    };
-
-    const FeaturedMedia = getFeaturedMedia(genericFields);
-    const Header = (
-      <PageHeader
-        breadcrumbs={breadcrumbs}
-        labels={{ labels: series.labels }}
-        title={series.title}
-        ContentTypeInfo={null}
-        Background={<HeaderBackground hasWobblyEdge={true} />}
-        FeaturedMedia={FeaturedMedia}
-        HeroPicture={null}
-      />
-    );
-
-    const upcomingEvents = events
-      .filter(event => {
-        const lastStartTime =
-          event.times.length > 0
-            ? event.times[event.times.length - 1].range.startDateTime
-            : null;
-        const inTheFuture = lastStartTime
-          ? new Date(lastStartTime) > new Date()
-          : false;
-        return inTheFuture;
-      })
-      .sort((a, b) => a.dateRange.firstDate - b.dateRange.firstDate);
-    const upcomingEventsIds = upcomingEvents.map(event => event.id);
-    const pastEvents = events
-      .filter(event => upcomingEventsIds.indexOf(event.id) === -1)
-      .sort((a, b) => b.dateRange.firstDate - a.dateRange.firstDate)
-      .slice(0, 3);
-
-    return (
-      <PageLayout
-        title={series.title}
-        description={series.metadataDescription || series.promoText || ''}
-        url={{ pathname: `/event-series/${series.id}` }}
-        jsonLd={events.map(eventLd)}
-        openGraphType={'website'}
-        siteSection={'whats-on'}
-        imageUrl={series.image && convertImageUri(series.image.contentUrl, 800)}
-        imageAltText={series.image && series.image.alt}
-        globalContextData={globalContextData}
-      >
-        <ContentPage
-          id={series.id}
-          Header={Header}
-          Body={<Body body={series.body} pageId={series.id} />}
-          contributorProps={{ contributors: series.contributors }}
-        >
-          {upcomingEvents.length > 0 && (
-            <SearchResults items={upcomingEvents} title={`What's next`} />
-          )}
-          {upcomingEvents.length === 0 && (
-            <h2 className="h2">
-              No events scheduled at the moment, check back soon…
-            </h2>
-          )}
-
-          {pastEvents.length > 0 && (
-            <Space v={{ size: 'xl', properties: ['margin-top'] }}>
-              <SearchResults
-                items={pastEvents}
-                title={`What we've done before`}
-              />
-            </Space>
-          )}
-        </ContentPage>
-      </PageLayout>
-    );
-  }
-}
-
-export default EventSeriesPage;
